fix: add @pulumi/cloudflare dependency and create separate stack definition
- Add @pulumi/cloudflare to infra/package.json devDependencies
- Create stack.ts with PagesDomain resource definition
- Simplify index.ts to use file-based stack instead of inline program
- Remove unused pulumi and cloudflare imports from index.ts
