+++ infra/package.json
@@ -6,6 +6,7 @@
     "dotenv": "^16"
   },
   "devDependencies": {
+    "@pulumi/cloudflare": "^5",
     "@pulumi/pulumi": "^3",
     "ts-node": "^10",
     "typescript": "^5"
